backend/requirements.txt
jupyter
notebook
polars
kagglehub
psycopg
psycopg[binary]
adbc-driver-postgresql
orjson
pydantic
fastapi
uvicorn
python-dotenv
rapidfuzz
spacy
langchain
langchain-groq
langchain_openai
langchain-google-genai
python-multipart
langchain-openrouter